feat(cluster): 实现基本的集群管理模式
- 添加集群通信和管理功能(注意尚未完善集群的通信协议，有很大BUG，会逐步修复)
- 实现节点发现、心跳机制和故障检测
- 添加槽位分配和键值路由功能
- 支持集群命令如 CLUSTER NODES、CLUSTER SLOTS 等

+++ Astra-CacheServer/proto/CommandImpl.hpp
@@ -52,6 +52,7 @@
 #pragma once
 #include "CommandResponseBuilder.hpp"
 #include "ICommand.hpp"
+#include "LuaExecutor.h"
 #include "caching/AstraCacheStrategy.hpp"
 #include "command_parser.hpp"
 #include "data/redis_types.hpp"
@@ -62,7 +63,6 @@
 #include <chrono>
 #include <datastructures/lru_cache.hpp>
 #include <memory>
-#include "LuaExecutor.h"
 
 namespace Astra::proto {
     using namespace datastructures;
@@ -1826,8 +1826,8 @@ namespace Astra::proto {
             size_t stop_idx = (stop >= 0) ? stop : (list_size + stop);
 
             // 确保索引在有效范围内
-            start_idx = std::max(start_idx, 0ULL);
-            stop_idx = std::min(stop_idx, list_size - 1);
+            start_idx = std::max(start_idx, static_cast<size_t>(0));
+            stop_idx = std::min(stop_idx, static_cast<size_t>(list_size - 1));
 
             // 如果范围无效，返回空数组
             if (start_idx > stop_idx || start_idx >= list_size) {
